+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 - [升级] 升级platform-bom版本到Cairo-RELEASE，支持Spring 5. [查看详情](https://docs.spring.io/platform/docs/Cairo-RELEASE/reference/htmlsingle/)
 - [升级] 升级Spring Cloud版本到Edgware.SR3
 - [升级] 升级Druid版到1.1.9，主要变更是支持Spring Boot 2.0
+- [优化] 优化xbird-starter-swagger2模块代码。
+- [优化] 优化xbird-starter-retrofit2模块代码。
 
-### 1.3.0
+### 1.3.0 2018-03-06
 
 - [新增] 集成HTTP客户端模块 `xbird-starter-retrofit2`。